UIDATIMP-7: Resolve PR comments :: Generalize tests interactors and Tags components :: DONE
