+++ src/components/ListColumnFormatters/TagsFormatter/index.js
@@ -1,0 +1,1 @@
+export * from './TagsFormatter';
