+++ src/components/ListColumnFormatters/index.js
@@ -1,2 +1,3 @@
+export * from './TagsFormatter';
 export * from './DateFormatter';
 export * from './UserNameFormatter';
