+++ src/settings/JobProfiles/JobProfiles.css
@@ -1,13 +1,3 @@
-.tags {
-  display: inline-flex;
-}
-
-.tagsIcon {
-  align-self: baseline;
-  margin-top: 0.3rem;
-  margin-right: 0.3rem;
-}
-
 .endOfRecord {
   text-align: left;
 }
