+++ src/utils/createOkapiHeaders.js
@@ -1,5 +1,5 @@
 /**
- * Builds OKAPI headers set.
+ * Builds OKAPI headers set
  *
  * @param {Object|okapi} okapi
  * @return {{'X-Okapi-Token': *, 'X-Okapi-Tenant': *}}
