+++ test/bigtest/interactors/match-profiles.js
@@ -1,28 +1,11 @@
 import {
   interactor,
   collection,
-  clickable,
-  blurrable,
-  value,
-  property,
 } from '@bigtest/interactor';
 
 import MultiColumnListInteractor from '@folio/stripes-components/lib/MultiColumnList/tests/interactor';
 
-@interactor class CheckboxInteractor {
-  static defaultScope = '[class*=checkbox---]';
-
-  clickLabel = clickable('label');
-  blurInput = blurrable('input');
-  inputValue = value('input');
-  isChecked = property('input', 'checked');
-
-  clickAndBlur() {
-    return this
-      .clickLabel()
-      .blurInput();
-  }
-}
+import { CheckboxInteractor } from './checkbox-interactor';
 
 @interactor class MatchProfilesInteractor {
   list = new MultiColumnListInteractor('#match-profiles-list');
